+++ src/entities/schedule/hooks/use-detailed-schedule.ts
@@ -1,0 +1,13 @@
+import { useSuspenseQuery } from '@tanstack/react-query'
+
+import { getDetailedSchedule } from '../models/detailed-schedule.API'
+import { scheduleQueryKeys } from '../models/schedule.query'
+
+export function useDetailedSchedule(scheduleId: string) {
+  const { data } = useSuspenseQuery({
+    queryKey: scheduleQueryKeys.detaildSchedule(scheduleId!),
+    queryFn: () => getDetailedSchedule(scheduleId),
+  })
+
+  return data
+}
