+++ src/entities/schedule/index.ts
@@ -1,2 +1,5 @@
 export * from './models/schedule.types'
 export * from './models/get-group-by-date'
+export * from './models/detailed-schedule.API'
+export * from './models/schedule.query'
+export * from './models/detailed-schedule.API'
