+++ src/entities/schedule/models/detailed-schedule.API.ts
@@ -2,7 +2,7 @@ import axiosInstance from '@/shared/api/axios-instance'
 
 import type { Schedule } from '@/entities/schedule'
 
-export default async function getDetailedSchedule(scheduleId: string): Promise<Schedule> {
+export async function getDetailedSchedule(scheduleId: string): Promise<Schedule> {
   const res = await axiosInstance(`/schedules/${scheduleId}`)
 
   return res.data.data
