+++ src/entities/schedule/models/schedule.query.ts
@@ -3,4 +3,6 @@
 export const scheduleQueryKeys = {
   all: ['schedules'] as const,
   userSchedules: (userId: number | string) => [...scheduleQueryKeys.all, 'user', userId] as const,
+  shareSchedules: (userId: number | string) => [...scheduleQueryKeys.all, 'share', userId] as const,
+  detaildSchedule: (scheduleId: string) => [...scheduleQueryKeys.all, 'schdule', scheduleId] as const,
 }
