+++ src/features/friend-schedule/hooks/use-friend-schedule.tsx
@@ -1,8 +1,7 @@
 import { useSuspenseQuery } from '@tanstack/react-query'
 import { useParams } from 'react-router'
 
-import groupByDate from '@/entities/schedule/models/get-group-by-date'
-import { scheduleQueryKeys } from '@/entities/schedule/models/schedule.query'
+import { groupByDate, scheduleQueryKeys } from '@/entities/schedule'
 
 import { getFriendSchedule } from '../api/friend-schedule.API'
