+++ src/features/share-schedule/hooks/use-share-schedule.ts
@@ -1,7 +1,7 @@
 import { useSuspenseQuery } from '@tanstack/react-query'
 import { useParams } from 'react-router'
 
-import groupByDate from '@/entities/schedule/models/get-group-by-date'
+import { groupByDate } from '@/entities/schedule'
 import { scheduleQueryKeys } from '@/entities/schedule/models/schedule.query'
 
 import { getShareSchedule } from '../api/share-schedule.API'
@@ -9,7 +9,7 @@ import { getShareSchedule } from '../api/share-schedule.API'
 export function useShareSchedule() {
   const { userId } = useParams<{ userId: string }>()
   const { data: schedules = [] } = useSuspenseQuery({
-    queryKey: scheduleQueryKeys.userSchedules(userId!),
+    queryKey: scheduleQueryKeys.shareSchedules(userId!),
     queryFn: () => getShareSchedule(userId!),
   })
